Fix review changes
